+++ packages/nimble-components/src/select/tests/select.react.tsx
@@ -1,4 +1,0 @@
-import { Select } from '..';
-import { wrap } from '../../utilities/tests/react-wrapper';
-
-export const NimbleSelect = wrap(Select);