+++ packages/nimble-components/src/tabs-toolbar/tests/tabs-toolbar.react.tsx
@@ -1,4 +1,0 @@
-import { TabsToolbar } from '..';
-import { wrap } from '../../utilities/tests/react-wrapper';
-
-export const NimbleTabsToolbar = wrap(TabsToolbar);