+++ packages/storybook/src/nimble/table-column/anchor/table-column-anchor.stories.ts
@@ -1,0 +1,137 @@
+import { html, ref } from '@microsoft/fast-element';
+import type { HtmlRenderer, Meta, StoryObj } from '@storybook/html';
+import { withActions } from '@storybook/addon-actions/decorator';
+import { tableTag } from '@ni/nimble-components/dist/esm/table';
+import { tableColumnTextTag } from '@ni/nimble-components/dist/esm/table-column/text';
+import { AnchorAppearance } from '@ni/nimble-components/dist/esm/anchor/types';
+import { tableColumnAnchorTag } from '@ni/nimble-components/dist/esm/table-column/anchor';
+import {
+    sharedTableActions,
+    SharedTableArgs,
+    sharedTableArgs,
+    sharedTableArgTypes
+} from '../base/table-column-stories-utils';
+import { createUserSelectedThemeStory } from '../../../utilities/storybook';
+
+const metadata: Meta<SharedTableArgs> = {
+    title: 'Components/Table Column: Anchor',
+    decorators: [withActions<HtmlRenderer>],
+    parameters: {
+        actions: {
+            handles: sharedTableActions
+        }
+    },
+    // prettier-ignore
+    argTypes: {
+        ...sharedTableArgTypes,
+        selectionMode: {
+            table: {
+                disable: true
+            }
+        },
+    }
+};
+
+export default metadata;
+
+const simpleData = [
+    {
+        firstName: 'Ralph',
+        lastName: 'Wiggum',
+        url: 'https://www.google.com/search?q=ralph+wiggum'
+    },
+    {
+        firstName: 'Milhouse',
+        lastName: 'Van Houten',
+        url: 'https://www.google.com/search?q=milhouse+van+houten'
+    },
+    {
+        firstName: 'Ned',
+        lastName: 'Flanders',
+        url: 'https://www.google.com/search?q=ned+flanders'
+    },
+    {
+        firstName: 'Maggie (no link)',
+        lastName: 'Simpson'
+    },
+    {
+        lastName: 'Simpson',
+        url: 'https://www.google.com/search?q=simpsons'
+    },
+    {
+        lastName: 'Simpson'
+    }
+] as const;
+
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+interface AnchorColumnTableArgs extends SharedTableArgs {
+    labelFieldName: string;
+    hrefFieldName: string;
+    appearance: keyof typeof AnchorAppearance;
+    underlineHidden: boolean;
+    placeholder: string;
+}
+
+export const anchorColumn: StoryObj<AnchorColumnTableArgs> = {
+    parameters: {},
+    // prettier-ignore
+    render: createUserSelectedThemeStory(html<AnchorColumnTableArgs>`
+        <${tableTag}
+            ${ref('tableRef')}
+            data-unused="${x => x.updateData(x)}"
+        >
+            <${tableColumnAnchorTag}
+                label-field-name="${x => x.labelFieldName}"
+                href-field-name="${x => x.hrefFieldName}"
+                appearance="${x => x.appearance}"
+                ?underline-hidden="${x => x.underlineHidden}"
+                placeholder="${x => x.placeholder}"
+            >
+            Link Column
+            </${tableColumnAnchorTag}>
+            <${tableColumnTextTag}
+                field-name="lastName"
+            >
+            Last Name
+            </${tableColumnTextTag}>
+        </${tableTag}>
+    `),
+    argTypes: {
+        labelFieldName: {
+            name: 'label-field-name',
+            description:
+                'Set this attribute to identify which field in the data record contains the visible text value for each cell in the column. The field values must be of type `string`. If a given row does not define a property with this name, that row will use the url as the label.',
+            options: ['firstName', 'lastName'],
+            control: { type: 'radio' }
+        },
+        hrefFieldName: {
+            name: 'href-field-name',
+            description:
+                'Set this attribute to identify which field in the data record contains the link url for each cell in the column. If the field is not defined in a particular record, that cell will be displayed as plain text instead of a link. The field values must be of type `string`.',
+            control: { type: 'none' }
+        },
+        appearance: {
+            options: Object.keys(AnchorAppearance),
+            control: { type: 'radio' },
+            description:
+                'Set to `prominent` to make the anchor appear in a different color than normal text.'
+        },
+        underlineHidden: {
+            name: 'underline-hidden',
+            defaultValue: { summary: 'false' },
+            description: 'Causes the underline to be hidden except on hover.'
+        },
+        placeholder: {
+            description:
+                'The placeholder text to display when the label and href are both `undefined` or `null` for a record.'
+        }
+    },
+    args: {
+        labelFieldName: 'firstName',
+        hrefFieldName: 'url',
+        appearance: 'default',
+        underlineHidden: false,
+        placeholder: 'Mystery',
+        ...sharedTableArgs(simpleData)
+    }
+};