+++ packages/storybook/src/nimble/wafer-map/wafer-map.react.tsx
@@ -1,0 +1,4 @@
+import { WaferMap } from '@ni/nimble-components/dist/esm/wafer-map';
+import { wrap } from '../../utilities/react-wrapper';
+
+export const NimbleWaferMap = wrap(WaferMap);